news-agent-backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
requests
feedparser
beautifulsoup4
lxml
resend